Split more 3D primitives into their own modules
